+++ slides-now-core.js
@@ -1,4 +1,4 @@
-/*! slides-now-core - 0.0.0 built on 2014-02-24
+/*! slides-now-core - 0.0.3 built on 2014-03-06
 author: Gleb Bahmutov <gleb.bahmutov@gmail.com>, support: @bahmutov */
 
 // Uses CSS to position elements in the center of its parent
@@ -113,7 +113,7 @@ window.recenterCodeBlocks = function (recenterOnResize) {
 (function (globals) {
     'use strict';
 
-    var messages, predicates, functions, verify, maybe;
+    var messages, predicates, functions, verify, maybe, not;
 
     predicates = {
         like: like,
@@ -162,11 +162,14 @@ window.recenterCodeBlocks = function (recenterOnResize) {
     functions = mixin(functions, predicates);
     verify = createModifiedPredicates(verifyModifier);
     maybe = createModifiedPredicates(maybeModifier);
+    not = createModifiedPredicates(notModifier);
     verify.maybe = createModifiedFunctions(verifyModifier, maybe);
+    verify.not = createModifiedFunctions(verifyModifier, not);
 
     exportFunctions(mixin(functions, {
         verify: verify,
-        maybe: maybe
+        maybe: maybe,
+        not: not
     }));
 
     /**
@@ -536,6 +539,17 @@ window.recenterCodeBlocks = function (recenterOnResize) {
         };
     }
 
+    /**
+     * Public modifier `not`.
+     *
+     * Negates `predicate`.
+     */
+    function notModifier (predicate) {
+        return function () {
+            return !predicate.apply(null, arguments);
+        };
+    }
+
     function createModifiedPredicates (modifier) {
         return createModifiedFunctions(modifier, predicates);
     }
@@ -610,7 +624,7 @@ block.list = replace(block.list)
 block._tag = '(?!(?:'
   + 'a|em|strong|small|s|cite|q|dfn|abbr|data|time|code'
   + '|var|samp|kbd|sub|sup|i|b|u|mark|ruby|rt|rp|bdi|bdo'
-  + '|span|br|wbr|ins|del|img)\\b)\\w+(?!:/|@)\\b';
+  + '|span|br|wbr|ins|del|img)\\b)\\w+(?!:/|[^\\w\\s@]*@)\\b';
 
 block.html = replace(block.html)
   ('comment', /<!--[\s\S]*?-->/)
@@ -1083,6 +1097,8 @@ function InlineLexer(links, options) {
   this.options = options || marked.defaults;
   this.links = links;
   this.rules = inline.normal;
+  this.renderer = this.options.renderer || new Renderer;
+  this.renderer.options = this.options;
 
   if (!this.links) {
     throw new
@@ -1146,11 +1162,7 @@ InlineLexer.prototype.output = function(src) {
         text = escape(cap[1]);
         href = text;
       }
-      out += '<a href="'
-        + href
-        + '">'
-        + text
-        + '</a>';
+      out += this.renderer.link(href, null, text);
       continue;
     }
 
@@ -1159,11 +1171,7 @@ InlineLexer.prototype.output = function(src) {
       src = src.substring(cap[0].length);
       text = escape(cap[1]);
       href = text;
-      out += '<a href="'
-        + href
-        + '">'
-        + text
-        + '</a>';
+      out += this.renderer.link(href, null, text);
       continue;
     }
 
@@ -1204,43 +1212,35 @@ InlineLexer.prototype.output = function(src) {
     // strong
     if (cap = this.rules.strong.exec(src)) {
       src = src.substring(cap[0].length);
-      out += '<strong>'
-        + this.output(cap[2] || cap[1])
-        + '</strong>';
+      out += this.renderer.strong(this.output(cap[2] || cap[1]));
       continue;
     }
 
     // em
     if (cap = this.rules.em.exec(src)) {
       src = src.substring(cap[0].length);
-      out += '<em>'
-        + this.output(cap[2] || cap[1])
-        + '</em>';
+      out += this.renderer.em(this.output(cap[2] || cap[1]));
       continue;
     }
 
     // code
     if (cap = this.rules.code.exec(src)) {
       src = src.substring(cap[0].length);
-      out += '<code>'
-        + escape(cap[2], true)
-        + '</code>';
+      out += this.renderer.codespan(escape(cap[2], true));
       continue;
     }
 
     // br
     if (cap = this.rules.br.exec(src)) {
       src = src.substring(cap[0].length);
-      out += '<br>';
+      out += this.renderer.br();
       continue;
     }
 
     // del (gfm)
     if (cap = this.rules.del.exec(src)) {
       src = src.substring(cap[0].length);
-      out += '<del>'
-        + this.output(cap[1])
-        + '</del>';
+      out += this.renderer.del(this.output(cap[1]));
       continue;
     }
 
@@ -1265,31 +1265,12 @@ InlineLexer.prototype.output = function(src) {
  */
 
 InlineLexer.prototype.outputLink = function(cap, link) {
-  if (cap[0].charAt(0) !== '!') {
-    return '<a href="'
-      + escape(link.href)
-      + '"'
-      + (link.title
-      ? ' title="'
-      + escape(link.title)
-      + '"'
-      : '')
-      + '>'
-      + this.output(cap[1])
-      + '</a>';
-  } else {
-    return '<img src="'
-      + escape(link.href)
-      + '" alt="'
-      + escape(cap[1])
-      + '"'
-      + (link.title
-      ? ' title="'
-      + escape(link.title)
-      + '"'
-      : '')
-      + '>';
-  }
+  var href = escape(link.href)
+    , title = link.title ? escape(link.title) : null;
+
+  return cap[0].charAt(0) !== '!'
+    ? this.renderer.link(href, title, this.output(cap[1]))
+    : this.renderer.image(href, title, escape(cap[1]));
 };
 
 /**
@@ -1335,6 +1316,149 @@ InlineLexer.prototype.mangle = function(text) {
 };
 
 /**
+ * Renderer
+ */
+
+function Renderer(options) {
+  this.options = options || {};
+}
+
+Renderer.prototype.code = function(code, lang, escaped) {
+  if (this.options.highlight) {
+    var out = this.options.highlight(code, lang);
+    if (out != null && out !== code) {
+      escaped = true;
+      code = out;
+    }
+  }
+
+  if (!lang) {
+    return '<pre><code>'
+      + (escaped ? code : escape(code, true))
+      + '\n</code></pre>';
+  }
+
+  return '<pre><code class="'
+    + this.options.langPrefix
+    + escape(lang, true)
+    + '">'
+    + (escaped ? code : escape(code, true))
+    + '\n</code></pre>\n';
+};
+
+Renderer.prototype.blockquote = function(quote) {
+  return '<blockquote>\n' + quote + '</blockquote>\n';
+};
+
+Renderer.prototype.html = function(html) {
+  return html;
+};
+
+Renderer.prototype.heading = function(text, level, raw) {
+  return '<h'
+    + level
+    + ' id="'
+    + this.options.headerPrefix
+    + raw.toLowerCase().replace(/[^\w]+/g, '-')
+    + '">'
+    + text
+    + '</h'
+    + level
+    + '>\n';
+};
+
+Renderer.prototype.hr = function() {
+  return '<hr>\n';
+};
+
+Renderer.prototype.list = function(body, ordered) {
+  var type = ordered ? 'ol' : 'ul';
+  return '<' + type + '>\n' + body + '</' + type + '>\n';
+};
+
+Renderer.prototype.listitem = function(text) {
+  return '<li>' + text + '</li>\n';
+};
+
+Renderer.prototype.paragraph = function(text) {
+  return '<p>' + text + '</p>\n';
+};
+
+Renderer.prototype.table = function(header, body) {
+  return '<table>\n'
+    + '<thead>\n'
+    + header
+    + '</thead>\n'
+    + '<tbody>\n'
+    + body
+    + '</tbody>\n'
+    + '</table>\n';
+};
+
+Renderer.prototype.tablerow = function(content) {
+  return '<tr>\n' + content + '</tr>\n';
+};
+
+Renderer.prototype.tablecell = function(content, flags) {
+  var type = flags.header ? 'th' : 'td';
+  var tag = flags.align
+    ? '<' + type + ' style="text-align:' + flags.align + '">'
+    : '<' + type + '>';
+  return tag + content + '</' + type + '>\n';
+};
+
+// span level renderer
+Renderer.prototype.strong = function(text) {
+  return '<strong>' + text + '</strong>';
+};
+
+Renderer.prototype.em = function(text) {
+  return '<em>' + text + '</em>';
+};
+
+Renderer.prototype.codespan = function(text) {
+  return '<code>' + text + '</code>';
+};
+
+Renderer.prototype.br = function() {
+  return '<br>';
+};
+
+Renderer.prototype.del = function(text) {
+  return '<del>' + text + '</del>';
+};
+
+Renderer.prototype.link = function(href, title, text) {
+  if (this.options.sanitize) {
+    try {
+      var prot = decodeURIComponent(unescape(href))
+        .replace(/[^\w:]/g, '')
+        .toLowerCase();
+    } catch (e) {
+      return '';
+    }
+    if (prot.indexOf('javascript:') === 0) {
+      return '';
+    }
+  }
+  var out = '<a href="' + href + '"';
+  if (title) {
+    out += ' title="' + title + '"';
+  }
+  out += '>' + text + '</a>';
+  return out;
+};
+
+Renderer.prototype.image = function(href, title, text) {
+  var out = '<img src="' + href + '" alt="' + text + '"';
+  if (title) {
+    out += ' title="' + title + '"';
+  }
+  out += '>';
+  return out;
+};
+
+/**
  * Parsing & Compiling
  */
 
@@ -1342,14 +1466,17 @@ function Parser(options) {
   this.tokens = [];
   this.token = null;
   this.options = options || marked.defaults;
+  this.options.renderer = this.options.renderer || new Renderer;
+  this.renderer = this.options.renderer;
+  this.renderer.options = this.options;
 }
 
 /**
  * Static Parse Method
  */
 
-Parser.parse = function(src, options) {
-  var parser = new Parser(options);
+Parser.parse = function(src, options, renderer) {
+  var parser = new Parser(options, renderer);
   return parser.parse(src);
 };
 
@@ -1358,7 +1485,7 @@ Parser.parse = function(src, options) {
  */
 
 Parser.prototype.parse = function(src) {
-  this.inline = new InlineLexer(src.links, this.options);
+  this.inline = new InlineLexer(src.links, this.options, this.renderer);
   this.tokens = src.reverse();
 
   var out = '';
@@ -1409,83 +1536,53 @@ Parser.prototype.tok = function() {
       return '';
     }
     case 'hr': {
-      return '<hr>\n';
+      return this.renderer.hr();
     }
     case 'heading': {
-      return '<h'
-        + this.token.depth
-        + ' id="'
-        + this.token.text.toLowerCase().replace(/[^\w]+/g, '-')
-        + '">'
-        + this.inline.output(this.token.text)
-        + '</h'
-        + this.token.depth
-        + '>\n';
+      return this.renderer.heading(
+        this.inline.output(this.token.text),
+        this.token.depth,
+        this.token.text);
     }
     case 'code': {
-      if (this.options.highlight) {
-        var code = this.options.highlight(this.token.text, this.token.lang);
-        if (code != null && code !== this.token.text) {
-          this.token.escaped = true;
-          this.token.text = code;
-        }
-      }
-
-      if (!this.token.escaped) {
-        this.token.text = escape(this.token.text, true);
-      }
-
-      return '<pre><code'
-        + (this.token.lang
-        ? ' class="'
-        + this.options.langPrefix
-        + this.token.lang
-        + '"'
-        : '')
-        + '>'
-        + this.token.text
-        + '</code></pre>\n';
+      return this.renderer.code(this.token.text,
+        this.token.lang,
+        this.token.escaped);
     }
     case 'table': {
-      var body = ''
-        , heading
+      var header = ''
+        , body = ''
         , i
         , row
         , cell
+        , flags
         , j;
 
       // header
-      body += '<thead>\n<tr>\n';
+      cell = '';
       for (i = 0; i < this.token.header.length; i++) {
-        heading = this.inline.output(this.token.header[i]);
-        body += '<th';
-        if (this.token.align[i]) {
-          body += ' style="text-align:' + this.token.align[i] + '"';
-        }
-        body += '>' + heading + '</th>\n';
+        flags = { header: true, align: this.token.align[i] };
+        cell += this.renderer.tablecell(
+          this.inline.output(this.token.header[i]),
+          { header: true, align: this.token.align[i] }
+        );
       }
-      body += '</tr>\n</thead>\n';
+      header += this.renderer.tablerow(cell);
 
-      // body
-      body += '<tbody>\n'
       for (i = 0; i < this.token.cells.length; i++) {
         row = this.token.cells[i];
-        body += '<tr>\n';
-        for (j = 0; j < row.length; j++) {
-          cell = this.inline.output(row[j]);
-          body += '<td';
-          if (this.token.align[j]) {
-            body += ' style="text-align:' + this.token.align[j] + '"';
-          }
-          body += '>' + cell + '</td>\n';
-        }
-        body += '</tr>\n';
-      }
-      body += '</tbody>\n';
 
-      return '<table>\n'
-        + body
-        + '</table>\n';
+        cell = '';
+        for (j = 0; j < row.length; j++) {
+          cell += this.renderer.tablecell(
+            this.inline.output(row[j]),
+            { header: false, align: this.token.align[j] }
+          );
+        }
+
+        body += this.renderer.tablerow(cell);
+      }
+      return this.renderer.table(header, body);
     }
     case 'blockquote_start': {
       var body = '';
@@ -1494,25 +1591,17 @@ Parser.prototype.tok = function() {
         body += this.tok();
       }
 
-      return '<blockquote>\n'
-        + body
-        + '</blockquote>\n';
+      return this.renderer.blockquote(body);
     }
     case 'list_start': {
-      var type = this.token.ordered ? 'ol' : 'ul'
-        , body = '';
+      var body = ''
+        , ordered = this.token.ordered;
 
       while (this.next().type !== 'list_end') {
         body += this.tok();
       }
 
-      return '<'
-        + type
-        + '>\n'
-        + body
-        + '</'
-        + type
-        + '>\n';
+      return this.renderer.list(body, ordered);
     }
     case 'list_item_start': {
       var body = '';
@@ -1523,9 +1612,7 @@ Parser.prototype.tok = function() {
           : this.tok();
       }
 
-      return '<li>'
-        + body
-        + '</li>\n';
+      return this.renderer.listitem(body);
     }
     case 'loose_item_start': {
       var body = '';
@@ -1534,24 +1621,19 @@ Parser.prototype.tok = function() {
         body += this.tok();
       }
 
-      return '<li>'
-        + body
-        + '</li>\n';
+      return this.renderer.listitem(body);
     }
     case 'html': {
-      return !this.token.pre && !this.options.pedantic
+      var html = !this.token.pre && !this.options.pedantic
         ? this.inline.output(this.token.text)
         : this.token.text;
+      return this.renderer.html(html);
     }
     case 'paragraph': {
-      return '<p>'
-        + this.inline.output(this.token.text)
-        + '</p>\n';
+      return this.renderer.paragraph(this.inline.output(this.token.text));
     }
     case 'text': {
-      return '<p>'
-        + this.parseText()
-        + '</p>\n';
+      return this.renderer.paragraph(this.parseText());
     }
   }
 };
@@ -1567,6 +1649,19 @@ function escape(html, encode) {
     .replace(/>/g, '&gt;')
     .replace(/"/g, '&quot;')
     .replace(/'/g, '&#39;');
+}
+
+function unescape(html) {
+  return html.replace(/&([#\w]+);/g, function(_, n) {
+    n = n.toLowerCase();
+    if (n === 'colon') return ':';
+    if (n.charAt(0) === '#') {
+      return n.charAt(1) === 'x'
+        ? String.fromCharCode(parseInt(n.substring(2), 16))
+        : String.fromCharCode(+n.substring(1));
+    }
+    return '';
+  });
 }
 
 function replace(regex, opt) {
@@ -1600,6 +1695,7 @@ function merge(obj) {
 
   return obj;
 }
+
 
 /**
  * Marked
@@ -1703,7 +1799,9 @@ marked.defaults = {
   silent: false,
   highlight: null,
   langPrefix: 'lang-',
-  smartypants: false
+  smartypants: false,
+  headerPrefix: '',
+  renderer: new Renderer
 };
 
 /**
@@ -1712,6 +1810,8 @@ marked.defaults = {
 
 marked.Parser = Parser;
 marked.parser = Parser.parse;
+
+marked.Renderer = Renderer;
 
 marked.Lexer = Lexer;
 marked.lexer = Lexer.lex;
@@ -2731,9 +2831,10 @@ var activeDeck, isPaused, p;
 
 p = window.progressFullWidth;
 
-p.bar = null;
-
-p.timerBar = null;
+if (p != null) {
+  p.bar = null;
+  p.timerBar = null;
+}
 
 activeDeck = null;
 
@@ -2741,6 +2842,9 @@ isPaused = false;
 
 bespoke.plugins.progressBar = function(deck) {
   var numberSlides;
+  if (p == null) {
+    return;
+  }
   if (p.bar != null) {
     p.bar.remove();
   }
@@ -2781,12 +2885,18 @@ bespoke.plugins.progressBar = function(deck) {
 };
 
 bespoke.plugins.progressBar.removeTimer = function() {
+  if (p == null) {
+    return;
+  }
   if (p.timerBar != null) {
     return p.timerBar.remove();
   }
 };
 
 bespoke.plugins.progressBar.timer = function(durationSeconds) {
+  if (p == null) {
+    return;
+  }
   if (p.timerBar != null) {
     p.timerBar.remove();
   }
@@ -2874,7 +2984,7 @@ directLinksToNewTab = function(html) {
 
 md2html = function(md) {
   var htmlParts, mdParts, severalBlankLines;
-  verify.string(md, 'expected markdown text');
+  verify.unemptyString(md, 'expected markdown text');
   severalBlankLines = /\n\r\n\r\n\r|\n\n\n|\n---/;
   mdParts = md.split(severalBlankLines);
   return htmlParts = mdParts.map(function(mdPart) {
@@ -2995,51 +3105,70 @@ var md2slides, optionsParser, verify;
 
 optionsParser = require('./options.coffee');
 
-require('./bespokeCounterPlugin.coffee');
+require('./bespoke-plugins/bespokeCounterPlugin.coffee');
 
-require('./bespokeShortcutPlugin.coffee');
+require('./bespoke-plugins/bespokeShortcutPlugin.coffee');
 
-require('./bespokeProgressBar.coffee');
+require('./bespoke-plugins/bespokeProgressBar.coffee');
 
-require('./bespokeThemePlugin.coffee');
+require('./bespoke-plugins/bespokeThemePlugin.coffee');
 
 md2slides = require('./md2slides.coffee');
 
 verify = require('check-types').verify;
 
-window.mdToPresentation = function(md, filename, element) {
-  var $article, addSlide, e, footerText, htmlParts, lastSlashAt, name, options, readable, wrapSection;
-  verify.unemptyString(md, 'expected markdown string');
-  if (element == null) {
-    element = $('div#dropzone');
+if (typeof $ === "undefined" || $ === null) {
+  throw new Error('Undefined jQuery $');
+}
+
+window.postProcessSlide = function($slide) {
+  var $img, caption;
+  $img = $('p > img', $slide);
+  if ($img.length === 1) {
+    $slide.empty().append($img);
+    caption = $img.attr('alt');
+    caption = caption.replace(/\ ?fullscreen$/, '');
+    if (caption) {
+      $slide.append('<p class="fullscreen-caption">' + caption + '</p>');
+    }
   }
-  verify.positiveNumber(element.length, 'invalid element to append to ' + element.selector);
+  return $slide;
+};
+
+window.mdToPresentation = function(opts) {
+  var $article, addSlide, e, footerText, htmlParts, lastSlashAt, md, name, options, readable, wrapSection;
+  if (opts == null) {
+    throw new Error('Missing presentatio options');
+  }
+  verify.unemptyString(opts.md, 'expected markdown string');
+  if (opts.element == null) {
+    throw new Error('Undefined element to bind presentation to');
+  }
+  verify.positiveNumber(opts.element.length, 'invalid element to append to ' + opts.element.selector);
   readable = window.innerWidth < 400;
   if (readable) {
     $('footer').text('');
-  } else if (filename) {
-    verify.unemptyString(filename, 'expected filename, got ' + filename);
-    name = filename;
-    lastSlashAt = filename.lastIndexOf('/');
+  } else if (opts.filename) {
+    verify.unemptyString(opts.filename, 'expected filename, got ' + opts.filename);
+    name = opts.filename;
+    lastSlashAt = opts.filename.lastIndexOf('/');
     if (lastSlashAt !== -1) {
-      name = filename.substr(lastSlashAt);
+      name = opts.filename.substr(lastSlashAt);
     }
     $('footer').text(name);
   }
   $('article.bespoke-parent').unbind();
   $('article').remove();
-  $article = element.append('<article>');
-  options = optionsParser.getSlidesNowOptions(md);
+  options = optionsParser.getSlidesNowOptions(opts.md);
+  md = optionsParser.removeOptionsLines(opts.md);
   if (readable) {
     $('body').removeClass('classic').addClass('full');
   } else if (options.theme != null) {
     verify.unemptyString(options.theme, 'expected string theme name ' + options.theme);
     $('body').removeClass('classic').addClass(options.theme);
   }
-  $('body').addClass('slides-now');
   footerText = options.footer || options.title;
   if ((footerText != null) && !readable) {
-    console.log('setting footer');
     $('footer').text(footerText);
   }
   if (options['font-family'] != null) {
@@ -3048,13 +3177,12 @@ window.mdToPresentation = function(md, filename, element) {
   if (options['font-size'] != null) {
     $('body').css('font-size', options['font-size']);
   }
-  md = optionsParser.removeOptionsLines(md);
   wrapSection = function(text) {
     var $slide;
     $slide = $('<section>\n' + text + '\n</section>\n');
     return $slide;
   };
-  $article = $('article');
+  $article = opts.element.append('<article>');
   addSlide = function(text) {
     var $slide, $span;
     if (text == null) {
@@ -3073,10 +3201,12 @@ window.mdToPresentation = function(md, filename, element) {
     } else {
       $slide = wrapSection(text);
     }
+    $slide = postProcessSlide($slide);
     return $('article').append($slide);
   };
   htmlParts = md2slides(md);
   htmlParts.forEach(addSlide);
+  $('body').addClass('slides-now');
   if (!readable) {
     try {
       if (options.timer != null) {
@@ -3087,8 +3217,12 @@ window.mdToPresentation = function(md, filename, element) {
     } catch (_error) {
       e = _error;
     }
-    recenter();
-    recenterImages();
+    if (opts.recenter) {
+      recenter();
+    }
+    if (opts.recenterImages) {
+      recenterImages();
+    }
     bespoke.horizontal.from('article', {
       hash: true,
       vertical: true,
@@ -3108,7 +3242,7 @@ window.mdToPresentation = function(md, filename, element) {
 };
 
 
-},{"./bespokeCounterPlugin.coffee":4,"./bespokeProgressBar.coffee":5,"./bespokeShortcutPlugin.coffee":6,"./bespokeThemePlugin.coffee":7,"./md2slides.coffee":8,"./options.coffee":9,"check-types":1}],11:[function(require,module,exports){
+},{"./bespoke-plugins/bespokeCounterPlugin.coffee":4,"./bespoke-plugins/bespokeProgressBar.coffee":5,"./bespoke-plugins/bespokeShortcutPlugin.coffee":6,"./bespoke-plugins/bespokeThemePlugin.coffee":7,"./md2slides.coffee":8,"./options.coffee":9,"check-types":1}],11:[function(require,module,exports){
 var S, removeSpeakerNotes, verify;
 
 verify = require('check-types').verify;
@@ -3139,4 +3273,4 @@ removeSpeakerNotes = function(md) {
 module.exports = removeSpeakerNotes;
 
 
-},{"check-types":1,"string":3}]},{},[4,5,6,7,8,9,10,11])
+},{"check-types":1,"string":3}]},{},[8,9,10,11])